u8 buff

+++ src/aligned.rs
@@ -109,7 +109,7 @@ impl Encode for Encoder {
         }
     }
 
-    fn close(self) -> (Box<[u64]>, u64) {
+    fn close(self) -> (Box<[u8]>, u64) {
         let mut this = self;
         this.w.write_bits(0xffff, 16);
         this.w.write_bit(0); // not sure why actual implementation does this
